defconfig: use cmdline console configuration for getty
Buildroot was configured to use the ttyS0 as the port for getty,
this created a problem for boards using the Marvell UART driver
where the device it ttyMV0.

To fix this BR2 can be configured to use default settings which
is to use the kernel cmdline (set by u-boot) console= parameter.
This way if boards uses the ttyX0, this parameter is passed to
kernel and is used by /etc/inittab getty respawn.

Change-Id: I682258e2cb17d8e860aba23334a8675365e3f1f8
Signed-off-by: Ofer Heifetz <oferh@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/28477
Tested-by: Star_Automation <star@marvell.com>